Substantial bug fixes.  Realizability checking finally works!
It has been tested on a variety of small to not-so-small gridworld
problems, for which realizability can be decided by visual inspection.

Be sure to set compile flags regarding word size (32 or 64 bit) to be
the same for gr1c and CUDD.  Specifically, define SIZEOF_LONG and
SIZEOF_VOID_P as appropriate. This affects how CUDD itself is built
and various definitions and macros in the CUDD header files included
by gr1c source code.