2-norm computation now can use all variables specified in a list
Try
  $ ./gr1c -l -m 20 examples/jit/tunnel.spc

Initial positions are still hard-coded in main.c; hard-coded positions
for gw2goals1obs.spc are in the comments next to those for tunnel.spc.

There is something strange occurring in the Min, Max JIT computation
for the first sublevel set.  In the present commit, it is explicitly
skipped during the horizon computation, at line 470 of solve_metric.c