compute_sublevel_sets() returns exactly the iterates
...rather than first modifying the smallest Y iterate to make it a set
of system goal states.  Providing access to the actual iterates (from
the vanilla fixed point formula) allows for a correct computation of
Min and Max values as used in the JIT algorithm.  As of this commit,
bounds_state() is hard-coded to use the 1-norm; see comments for the
2-norm.

Beware that initial values for simulation (flag -m at the
command-line) are still hard-coded in main.c