Improve simulation configuration from command-line arguments
See the help listing; try invoking gr1c with "-h".  For examples, try
  $ ./gr1c -v -m -1,"x y" examples/jit/tunnel.spc
  $ ./gr1c -v -m 30,"x y","1 2 15 7" examples/jit/tunnel.spc
  $ ./gr1c -v -m 30,"x y","0 0 7 2" examples/jit/gw2goals1obs.spc

In particular, variables to include in the 1-norm computation ("metric
variables") can be given at the command-line, along with the initial
state of a simulation.  This has not been tested on specifications
involving mixed boolean and (bounded) integer variables.

There are now two verbosity levels: "-v" for "verbose" and "-vv" for
"very verbose."