Implementation of verification model generation for Spin
The command-line flag for gr1c is -P.  It may eventually support
selection of the tool target, but will continue with present target
(Spin) as default.

Consult documentation for spin_aut_dump() (cf. src/automaton.h).
Note that the resulting Promela file is missing a never claim that
may be added using the formula near the top of the file (under "LTL
formula").  E.g., try

  $ ./gr1c -t txt -o foo.txt examples/trivial.spc -P > trivsoln.pml
  $ spin -f '!(...formula inside comments of trivsoln.pml...)' >> trivsoln.pml
  $ spin -a trivsoln.pml && gcc -DNOREDUCE -o pan pan.c && ./pan -a

Notice that the formula must be negated to create a never claim.
We then use Spin to try to find acceptance cycles, i.e., cycles in
which the negated formula is satisfied, i.e., cycles in which the
specification is violated.