Remove assumption of uniqueness of nodes w.r.t. state and mode
This is a cross-cutting change that provides a more general representation
of strategies.  It should also make debugging easier because the output
routines previously made that assumption.  A new function, anode_index(),
provides generation of node indices using anode_t pointers.

Finally, grjit and sim_rhc() have been marked in the corresponding
documentation with warnings that they were not thoroughly checked as
part of this change, and so should be used with caution.