Correct identifier regex that defines input syntax
The regular expression that defines valid identifiers (variable names)
in the input syntax of gr1c had an error (or feature) that permitted
the following to be used after the first letter: [\]^`

Only two of these, [ and ], are used at all in the current syntax.
Furthermore, allowing [ or ] to be used in identifiers after the first
letter does not cause ambiguities in the grammar because [ and ] are
only used to declare a range of an integer domain or to represent the
"always" (G) temporal operator in the originally intended syntax.
Thus, rather than an error, we may regard this as an unintended
feature that is now obsolete and removed without warning.

I give a regression test that is motivated by the confusing situation
that led to this discovery.