Variety of exit codes; prefer exit(-1) following perror()
Note the new behavior for the file format version switch of `autman`.
`autman` now prints the number to stdout and returns 0 (success),
instead of returning the version number as the exit code.

Besides changing return values, in this changeset documentation is
updated accordingly. The main description about exit codes is on the
introductory page of the manual.