Fix several non-critical memory leaks, found by cppcheck
The memory leaks are non-critical because they were code paths that
should quickly lead to program termination, so the memory would soon
be garbage-collected anyway.

This changeset also includes a minor performance improvement to not
perform a dead assignment to a function parameter in the
implementation of spin_aut_dump().

Websites about cppcheck are
http://cppcheck.sourceforge.net
https://github.com/danmar/cppcheck/
https://sourceforge.net/projects/cppcheck/