Initialize pointer with NULL, just in case
There is no currently known code path that would lead to ptree_BDD()
returning `fn` without having assigned to it. However, following the
recommendation of the Clang Static Analyzer, here it is initialized
with NULL to be sure.

It is missing from the commit message of de76344c58cfeeec5a714a41d73bbfbc23e1dbd4,
but note that in commit de76344c58cfeeec5a714a41d73bbfbc23e1dbd4
the allocator size for X_ijr in synthesize() was changed from
sizeof(DdNode *) to sizeof(DdNode **). Correctness would not be
affected because the result is an array of pointers, so alignment and
stride are the same.