Check for empty calls to rm_sysgoal()
This can occur if rm_sysgoal() (prototype in patching.h) is called
with a strategy that has neither nodes marked explicitly with the mode
to be deleted nor nodes marked with the next mode (modulo the number
of liveness goals).

Before this changeset, subsequent errors could occur in this case.
E.g., the potential error found by the Clang static analyzer is calls
to malloc() for size 0.