Free memory to avoid leaks
In the current implementation, the case of the leaks from main() is
not important because the memory would be recovered anyway after
program termination.

The leak from check_realizable_internal() occurs as part of returning
with error after deciding that a parameter has an unexpected value.
Though not demonstrated in the current implementation, we can imagine
future designs or other uses of code from gr1c where this leak matters.

Both were found by the Clang static analyzer.