Add build root as first search path for test-cli
Instead of making it the only entry in PATH, it is enough to have it
first. Future additions to test-cli.sh might call other programs on
the system path. Besides, adding a path for gr1c to the existing list
of search paths is likely how the user will configure her shell.