Add test cases that involve non-existing files
To do this, a format is introduced in test-cli.sh to check for already
existing files in the tests/ directory.

TEST CASE: gr1c -- rg

which is motivated because `gr1c rg` is valid (the `rg` command),
whereas `gr1c -- rg` is for a file named "rg".

TEST CASE: gr1c f00f

should fail because the file named "f00f" does not exist.