Option to redirect LTL formula for Spin Promela
Before this changeset, the LTL formula available for model checking
strategies with Spin was only included in the comments of the
generated Promela model. Using it required manual intervention (or
fragile string parsing).

From this changeset, spin_aut_dump() can be given a file pointer by
which to print the LTL formula. It is demonstrated in `gr1c autman`,
where the -o and -P switches together cause the Promela model to be
written to a file and the LTL formula to be printed to stdout. This
usage of `gr1c autman` is demonstated in test-verification.sh.

Note that the `-o` switch of autman that is introduced here is more
general than code generation for model checkers. It can redirect other
output from `gr1c autman`, e.g., in conversion among strategy formats.