Simplify conditions of formula_fp in spin_aut_dump
This changeset fixes a bug in assigning default values to formula_fp
in spin_aut_dump(). Before this changeset, if both fp and formula_fp
were assigned stdout at invocation of spin_aut_dump(), then the
formula was printed twice.

The conditional statements for assigning values are simplified now.
Though there might expressions that ensure formula_fp is not assigned
values twice, the code in this changeset is preferred for ease of
understanding and maintenance.