Update regression test given previous commit
Before this changeset, the expected output did not mark one of the
nodes as initial, despite it being so. Now it is marked initial, so
this changeset updates the expected output that is used for testing.

As described in the User's manual and API documentation, a node with
initial field `false` can still be viable as an initial node for
execution. In other words, initial `true` is sufficient but not
necessary for plays from a node to satisfy initial conditions.
Therefore, both the expected outputs from before and after this
changeset are correct.

Finally, note that `initial` may eventually be used as a sufficient
and necessary indication.