TEST: Add more model checking tests using Spin
The additional files are from the list in test-gr1c.sh used for
checking cases that are expected to be realizable. The only file from
that list not included in test-verification.sh is arbiter4.spc, which
has a solution that involves a relatively large duration to check.

With this changeset, the list `REFSPECS` in test-verification.sh is
now sorted alphabetically.