BUILD: Move "exp" and "aux" into `make core`
The commands `patch` and `autman` have been promoted to being a part
of the core configuration, so now it suffices to use `make` to build
them.

This change is motivated in part in anticipation of new features,
e.g., an HTTP server, that will involve dependencies that some users
will not want to satisfy for basic operations. The intent is that
`make` should provide programs for the current features. The target
`make all` will be used for the forthcoming features.