Allow ENV:...; and others to have blank space prefix
This is motivated to support terminals that use \n\r for line-breaks.
Before this changeset, it was a syntax error for ENV, SYS, SYSINIT,
and others to have any space between them and the start of the line.
With this changeset, in terms of the lexer, SYS: x;\n\rSYSGOAL: []<>x;
appears as

    SYS: x;
    \rSYSGOAL: []<>x;

and `SYSGOAL` is recognized because the prefix `\r` is "blank space".

Note that \r does not contribute to the line count, which may cause
trouble in situations where it renders as a line-break in addition to
any line-breaks from \n characters.