DOC: Stop using `$` prefix for CLI commands
Most examples of commands to be entered at a terminal (also called
shell or command-line interface (CLI)) begin with `$` as a matter of
style to remind the reader about where to try the command and for
emphasis that superuser (root) is not required. However, in this
changeset, this prefix is removed because `sudo` can simply be used
instead of changing the prefix to `#`, and the shell prefix `$`
interferes with copy-and-paste usage of examples.