DOC: Note append_anode_trans does not check dup
Note in the API documentation of append_anode_trans() that it does not
check for duplicate outgoing edges.

While there is some risk of creating malformed strategy automata, this
behavior is nonetheless motivated for efficiency.

Also, correct return-value description, which is simply `head` if
success. It can never be a new `head` pointer because the function
never modifies `head`.