Only fetch dependency tar file if one not found
get-deps.sh will download the CUDD distribution tarball only if a file
of the same name does not already exist locally. In any case, the SHA
hash of the local file is checked. This change is motivated by
avoiding unnecessary network activity and accidental deletion of your
local copy of the file if `curl` fails.

Changes and motivations for get-extra-deps.sh are entirely similar.