predicate macros for building on Windows
With this changeset, building gr1c for Windows is operational but not
fully tested. To do it, first install a cross-compiler from the
mingw-w64 project (http://mingw-w64.org/). Major GNU/Linux
distributions have this available in respective package repositories;
e.g., on Fedora or RedHat, `dnf install mingw64-gcc`, and on Ubuntu,
`apt-get install gcc-mingw-w64`. Then,

    export CC=x86_64-w64-mingw32-gcc
    export LD="x86_64-w64-mingw32-ld -r"

Then, for building CUDD,

    ./configure --prefix=`pwd`/../.. --host=x86_64-w64-mingw32

and for gr1c,

    make -e

This is part of work for issue #34
and https://github.com/tulip-control/tulip-control/issues/197

* about _WIN32 and _WIN64 macros, https://docs.microsoft.com/en-us/cpp/preprocessor/predefined-macros?view=vs-2019
* about _spawnvp(), https://docs.microsoft.com/en-us/cpp/c-runtime-library/spawn-wspawn-functions?view=vs-2019